requirements.txt
biopython>=1.81
requests>=2.31.0
certifi>=2023.0.0
matplotlib>=3.7.0
reportlab>=4.0.0
